recovery: add GSO-friendly pacer
Currently we update the timestamp for next packet
(`last_packet_scheduled_time`) for each packet based on
the current pacing rate, which can confuse the
rtt measurement because the recorded timestamp
and actual time sent can be different, especially when
the application sends in burst using send_quantum
and batches a group packets using GSO. In such case
rtt measured can be smaller than expected, or even 0
when the packet is not sent in a specified time.

Pacer provides the timestamp for next packet to be sent
based on the current send_quantum, pacing rate
and last updated time. It is reset when a valid packet is
received (including ACK and non-ACK packets) to
start a new burst.

Pacer is GSO-friendly, which means it will provides a same
timestamp for same sized (except last one) packets in a
GSO group.